requirements.txt
Flask==3.0.2
google.cloud.bigquery
google.cloud
